Move Filename and YamlFile to root of namespace
