+++ src/Collection/Deserializer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace SmartAssert\YamlFile\Collection;
 
-use SmartAssert\YamlFile\Model\YamlFile;
+use SmartAssert\YamlFile\YamlFile;
 use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Parser;
