+++ src/Validator/YamlFileValidator.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace SmartAssert\YamlFile\Validator;
 
-use SmartAssert\YamlFile\Model\YamlFile;
 use SmartAssert\YamlFile\Validation\Validation;
 use SmartAssert\YamlFile\Validation\ValidationInterface;
 use SmartAssert\YamlFile\Validation\YamlFileContext;
+use SmartAssert\YamlFile\YamlFile;
 
 class YamlFileValidator
 {
