+++ src/Validator/YamlFilenameValidator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace SmartAssert\YamlFile\Validator;
 
-use SmartAssert\YamlFile\Model\Filename;
+use SmartAssert\YamlFile\Filename;
 use SmartAssert\YamlFile\Validation\FilenameContext;
 use SmartAssert\YamlFile\Validation\Validation;
 use SmartAssert\YamlFile\Validation\ValidationInterface;
