+++ src/YamlFile.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\YamlFile\Model;
+namespace SmartAssert\YamlFile;
 
 class YamlFile
 {
