+++ tests/Unit/Validator/YamlFileValidatorTest.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace SmartAssert\Tests\YamlFile\Unit\Validator;
 
 use PHPUnit\Framework\TestCase;
-use SmartAssert\YamlFile\Model\Filename;
-use SmartAssert\YamlFile\Model\YamlFile;
+use SmartAssert\YamlFile\Filename;
 use SmartAssert\YamlFile\Validation\ContentContext;
 use SmartAssert\YamlFile\Validation\FilenameContext;
 use SmartAssert\YamlFile\Validation\Validation;
@@ -15,6 +14,7 @@ use SmartAssert\YamlFile\Validation\YamlFileContext;
 use SmartAssert\YamlFile\Validator\ContentValidator;
 use SmartAssert\YamlFile\Validator\YamlFilenameValidator;
 use SmartAssert\YamlFile\Validator\YamlFileValidator;
+use SmartAssert\YamlFile\YamlFile;
 use Symfony\Component\Yaml\Parser;
 
 class YamlFileValidatorTest extends TestCase
